Break up scheduler test file
This also cleans up some small lint issues
